# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VyOS API",
    "description": "REST proxy for one or more VyOS router devices. Exposes CRUD operations over interfaces, VRFs, VLANs, firewall policies, and firewall address groups. Each device is identified by the name supplied in the VYOS_HOSTS environment variable.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/devices/{device_id}/networks": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" }
      ],
      "get": {
        "tags": ["networks"],
        "summary": "List all interfaces",
        "description": "Retrieves the full `interfaces` config tree and returns every interface of every type with its IPv4 addresses.",
        "operationId": "listNetworks",
        "responses": {
          "200": {
            "description": "Interface list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/NetworkInfo" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "post": {
        "tags": ["networks"],
        "summary": "Set an interface address",
        "description": "Adds an IPv4 address (CIDR notation) to the specified interface, creating the interface config node if it does not exist.",
        "operationId": "createNetwork",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateNetworkRequest" },
              "example": {
                "interface": "eth0",
                "type": "ethernet",
                "address": "192.168.1.1/24",
                "description": "LAN"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interface address set",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NetworkInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/networks/{interface}": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/interface" },
        { "$ref": "#/components/parameters/iface_type_query" }
      ],
      "get": {
        "tags": ["networks"],
        "summary": "Get an interface",
        "description": "Retrieves the config for a single interface. Use the `type` query parameter when the interface is not ethernet.",
        "operationId": "getNetwork",
        "responses": {
          "200": {
            "description": "Interface details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NetworkInfo" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "put": {
        "tags": ["networks"],
        "summary": "Replace interface address",
        "description": "Deletes all existing addresses on the interface and sets the new one. `type` must be provided in the request body.",
        "operationId": "updateNetwork",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateNetworkRequest" },
              "example": {
                "type": "ethernet",
                "address": "10.0.0.1/24",
                "description": "uplink"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interface updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NetworkInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "delete": {
        "tags": ["networks"],
        "summary": "Delete an interface",
        "description": "Deletes the entire interface config node. Defaults to `type=ethernet`.",
        "operationId": "deleteNetwork",
        "responses": {
          "204": { "description": "Interface deleted" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "device_id": {
        "name": "device_id",
        "in": "path",
        "required": true,
        "description": "Device name as registered in VYOS_HOSTS",
        "schema": { "type": "string", "example": "router1" }
      },
      "interface": {
        "name": "interface",
        "in": "path",
        "required": true,
        "description": "Interface name (e.g. `eth0`, `bond0`, `lo`)",
        "schema": { "type": "string", "example": "eth0" }
      },
...
      "iface_type_query": {
        "name": "type",
        "in": "query",
        "required": false,
        "description": "Interface type in the VyOS config tree. Defaults to `ethernet`.",
        "schema": {
          "type": "string",
          "default": "ethernet",
          "enum": ["ethernet", "loopback", "bonding", "bridge", "wireguard", "dummy"]
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "device not found: router99" }
        }
      },
...
      "NetworkInfo": {
        "type": "object",
        "required": ["interface", "type", "addresses"],
        "properties": {
          "interface":   { "type": "string", "example": "eth0" },
          "type":        { "type": "string", "example": "ethernet" },
          "addresses":   { "type": "array", "items": { "type": "string" }, "example": ["192.168.1.1/24"] },
          "description": { "type": "string", "example": "LAN" }
        }
      },
      "CreateNetworkRequest": {
        "type": "object",
        "required": ["interface", "type", "address"],
        "properties": {
          "interface":   { "type": "string", "description": "Interface name", "example": "eth0" },
          "type":        { "type": "string", "description": "Interface type in the VyOS config tree", "example": "ethernet" },
          "address":     { "type": "string", "description": "IPv4 address in CIDR notation", "example": "192.168.1.1/24" },
          "description": { "type": "string", "description": "Alphanumeric label (no spaces)", "example": "LAN" }
        }
      },
      "UpdateNetworkRequest": {
        "type": "object",
        "required": ["type", "address"],
        "properties": {
          "type":        { "type": "string", "description": "Interface type in the VyOS config tree", "example": "ethernet" },
          "address":     { "type": "string", "description": "Replacement IPv4 address in CIDR notation", "example": "10.0.0.1/24" },
          "description": { "type": "string", "example": "uplink" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Missing or invalid request fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface, type, and address are required" }
          }
        }
      },
      "NotFound": {
        "description": "Device or resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface not found" }
          }
        }
      },
      "DeviceNotFound": {
        "description": "The requested device_id is not registered",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device not found: router99" }
          }
        }
      },
      "DeviceRejected": {
        "description": "The VyOS device rejected the operation (invalid config, constraint violation, etc.)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device rejected operation: Configuration path already exists" }
          }
        }
      },
      "DeviceError": {
        "description": "Could not communicate with the VyOS device (network error, TLS failure, timeout)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device communication error: dial tcp 192.168.1.1:443: connection refused" }
          }
        }
      }
    }
  }
}